Fix bug on click event position when canvas is scrolled
Observed behavior when the canvas was scrolled to the bottom
where wells at the top of the canvas would be getting clicked.
This is because event.x and event.y reference coordinates relative
to the current view of the canvas, but were passed to a function
that accepts absolute canvas coordinates. Solved by converting
back using the canvasx and canvasy methods, per this StackExchange:

https://stackoverflow.com/a/59619511