Handle excel files where not all wells are included
Per discussion with Juanita, it is not always the case
that all 96 wells are included in the file generated, and that
the WellNum labeling counts what is included in the file, not
absolute plate coordinates.

The new behavior should be that we wait to write the condition
information to the pandas dataframe until we decide to save, at
which point the well number is inferred from the wells that have
been selected to have labeled conditions.