Added main analysis script
This main.py script is a minimally modified version of the script
used for the 08_16_20 analysis with several changes

- All filepaths constructed explicitly using / delimiters have been
changed to use os.path.join() instead, for cross-platform compatibility.
This was necessary since I am working on MacOS, which has a different
filepath convention.
- Additionally, all other string construction using manual concatenation
was switched to using f-strings. This was not necessary but does make
the code clearer and more concise.
- All hardcoded drug names were removed from the file to make it safe
to check into git. At the moment, those are provided through the command
line, but soon will be able to go through the UI.