Deprecate percents in main.py
In a previous commit, we deprecated percents in the plate annotation
UI since it is not strictly necessary (any information on % FBS can
just be rolled into the condition field). This commit takes out
all of script components in main.py that reference percents.
It also removes some hardcoded references to 10% FBS concentration
for consistency.