Refactor main.py to be a giant function
In order to create a plot UI, we will need some function
to call for plotting rather than the script as it exists.
I opted to turn main.py into one giant function rather than,
say, a class with different methods for different stats/plots
strategies, because that is the easiest way to make marginal
progress without breaking anything. A full refactoring into
a class may be desirable later.

This also includes some cleanup:
- Deprecating manual input of frames and conditions in favor of
specifying filenames and inferring that information from the data.
- Adding checks that raise errors if certain input options are invalid
- Changing boolean options from 0/1 encoding to True/False encoding
- Removing load_stats and do_chi2 options, which are not implemented