Infer base_directory from filepaths
The original draft of the Plot UI required manually inputting the
base_directory in addition to the csv filepaths for passing along
to the main.py script. This commit changes the main.py script to
infer the base_directory assuming that the files provided are in
the same folder. If they are not, it throws an error.