Add ability to annotate multiple frames at same time
I previously had the annotation UI assume that you are only annotating
one frame at a time, but this doesn't make sense, since all the annotations
should be the same for every frame.

This change deprecates the previous frame assignment feature in favor of
a series of prompts when you open the UI to specify the frame of each
file you have opened.