Calculate and save all normalizations
The current strategy for normalized values is to compute just the normalization
we care about for just the data we care about at each iteration of a loop through
some stats or plots. This commit changes the strategy to calculate every kind of
normalization for every row in advance of doing any plotting, and then writing
those values back out to the CSV. This has the advantage of removing duplicate
code.

Note that at the moment this has the disadvantage that all the normalizations
are computed for all values every time we run the plots, when theoretically
we could check whether the values have already been computed. This could be
changed in the future. This commit also only makes use of the precomputed
normalized values for lineplots. A future change will update the rest of
norm references once we ensure that the lineplots change works correctly.

This change also has the side effect of fixing some pandas warnings about assigning
values to copies of dataframes. Those disappear for the