Fix ttest-related bugs
- There was a variable (mi_t) that was undefined, replaced with actual reference (mi).
- The variable "l" was supposed to be a numerical index of the condition (i)
we were looking at, but the iteration was moved to the wrong loop when
refactoring. Fixed for now by using enumerate(), should look into completely
deprecating reliance on that index.
- There was an unnecessary SettingWithCopyWarning from coercing a column that was
already numeric to be numeric. If we really need that it should go in DataAnnotation
when the CSV is loaded.