Deprecate use of 'norm' in favor of norm_colname
In a previous commit, we refactored the normalization to go at the beginning
and then carry through the reference to which data should be used in the norm_colname
variable. This commit finishes the job by cleaning up all the places where
the old pattern was used. There are still two places where we explicitly check
the normalization_type in the ttest code. Those should probably be removed.

Additionally, I noticed in testing that it is not possible to run a Wilcoxon
test against the control when the data is control normalized, since all the values
for the control will simply equal 1, and that breaks the non-parametric test
(statsmodels throws an error). Added in an explicit warning at the beginning to
prevent this.