Add backfill script for adding a Date column to CSVs
This is (ideally) a one-time backfill script for the purpose of pulling the date information
from the folder that a csv is in and adding it as a column to the actual csv, for the purpose
of making it easier to use downstream when we plot multiple experiments at the same time.

Ran it on everything in the top-level OneDrive folder as well as everything in
all_data_combined, on 2021-03-08.