Remove references to 'total' normalization type
In the ttest code, there was a block that referred to total normalization.
The issue there was that in other normalization (such as relative), you
could end up dividing by zero. Change this to instead explicitly check
for dividing by zero and default to zero (defaulting to zero is consistent
with prior behavior).