Made left bar frame into a scrolling frame
We need to be able to scroll down on the left frame when there are
more conditions than will fit on the screen.

Giving a scrollbar to a frame is surprisingly difficult! You have to
basically trick tkinter into thinking the frame is actually a canvas.
Tried this for a while, including modifying some stackexchange code,
but nothing seemed to work correctly. Instead using this pip package.
It requires also installing the package through pip, but fortunately
you can specify pip installs in the conda environment.yml