allowed the automatic reading in for files named Drug_List.csv and Drug_List_my_order.csv for ordering plot columns
allowed the automatic reading in for files named Drug_List.csv and Drug_List_my_order.csv, which will determine the list of conditions to use for stats and plots. If these files are not present in the directory, it will simply proceed as usual using the conditions you provide in the gui and arrange them by mean value in the boxplots. both files are essentially just a sheet with one column with the header "Condition" and the list, in order, of the drugs you want. easily manually generated by taking the processed data sheet, copying over the Condition column, and run remove duplicates on it in excel under Data >> Tools, (for ease then sort alphabetically,) then arrange in list you want. If only Drug_list.csv is present, it will use those drugs but still sort by means. If you want them sorted by your order, just copy paste that file and rename it to  Drug_list_my_order.csv